7.9.audio_classification/requirements.txt
pydub
Tensorflow
matplotlib.pyplot